requirements/dev.txt
-r common.txt
forgerypy-0.1
Pygments-2.2.0 
certifi-2017.7.27.1 
chardet-3.0.4 
httpie-0.9.9 
idna-2.6 
requests-2.18.4 
urllib3-1.22
coverage-4.4.1
selenium-3.5.0
